Build SQLite ModelDB properly; build test dataset
Now the SQLite version of ModelDB builds properly.
Additionally, added build-test-db script that generates
a test SQLite database for use with unit testing.

Both databases include the USERS and UserSession tables
needed for ModelSEED and ModelSeedApi.

Neither script is intended to be called by regular
users. Both are used in cronjobs on the data server
to periodically rebuild the data-downloads for users.