Fix PPO and FIGMODELdatabase to work with SQLite
Previously, PPO would attempt to load DBD::mysql due
to the way that arguments to DBMaster were parsed.
Since installing DBD::mysql requires a working install
of MySQL on the machine, this was no good.

Now PPO accepts the required parameters for SQLite,
avoiding having to require DBD::mysql.

Added configuration to FIGMODELdatabase to properly
call DBMaster with the -backend argument for SQLite.