Fix unit tests - 3 errors remain
A nearly complete fix of all unit tests
to work with new deployment system. Two
outstanding errors in prove that are
visible and another commented out of
test scripts.

Remaining hidden error with rxnmdl
data that isn't getting copied into
TestDB.tgz, need to get models in
good state on "master" core.

Rewrite of TestingHelpers routines to
use TestDB.tgz which will be downloaded
from the server and unpacked in the
database root directory into TestData/

This directory is then copied to a temporary
directory for each newDebugFIGMODEL() call
from TestingHelpers. It is pretty efficient
compared with the old routine.